+++ MordoEngine/src/Scene/Editor/EditorScene.h
@@ -9,12 +9,13 @@
 class EditorScene : public Scene
 {
 private:
-	std::shared_ptr<Camera> m_Camera;
-	std::shared_ptr<Geomipmapping> m_Renderer;
-	std::unique_ptr<AreaSelectorRenderer> m_TerrainSelector;
 	std::shared_ptr<terrain::Terrain> m_Terrain;
+	std::shared_ptr<Camera> m_Camera;
+	std::shared_ptr<Renderer> m_Renderer;
+	std::unique_ptr<AreaSelectorRenderer> m_TerrainSelector;
+
 public:
-	EditorScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Geomipmapping> renderer , std::shared_ptr<Camera> camera);
+	EditorScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer);
 	void Render() override;
 	void Update(float deltaTime) override;
 	~EditorScene() override;