+++ MordoEngine/src/Scene/Game/GameScene.cpp
@@ -1,31 +1,12 @@
 #include "GameScene.h"
-#include "../../Terrain/HeightmapTerrain.h"
-#include "../../Terrain/FaultFormationTerrain.h"
-#include "../../Terrain/MidpointDisplacement.h"
-#include "../../API/OpenGL/OpenGLBackend.h"
-#include "../../Renderer/TriangleRenderer.h"
-#include "../../Renderer/Geomipmapping.h"
 #include "../../Core/Managers/Manager.h"
 #include "../../Core/Shader/Shader.h"
 #include "../../Input/Input.h"
 
-GameScene::GameScene() : Scene("terrain"),
-						//m_Terrain(std::make_shared<HeightMapTerrain>("res/maps/heightmap.raw")),
-						m_Terrain(std::make_shared<FaultFormationTerrain>(513, 3.0f, 50, 0, terrain::RAW_HEIGHT_MAX, 0.15f)),
-						//m_Terrain(std::make_unique<MidpointDisplacement>(1057, 3.0f, 1.0f, 0, terrain::RAW_HEIGHT_MAX)),
-						m_Sun(std::make_unique<Sun>(0.9f)),
-						m_Camera(std::make_shared<Camera>(glm::vec3{ 0.0f, 50.0f, .80f },
-							OpenGLBackend::SCR_WIDTH,
-							OpenGLBackend::SCR_HEIGHT))
+GameScene::GameScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer) 
+	: Scene("terrain"), m_Terrain(terrain) , m_Camera(camera) , m_Renderer(renderer) , m_Sun(std::make_unique<Sun>(0.9f))
 {
-	m_Terrain->SetHeightScale(200.0f * m_Terrain->GetWorldScale());
-	m_Camera->SetPosition(glm::vec3(m_Terrain->GetSize()/2,
-								   m_Terrain->GetHeightScale(), 
-		                           m_Terrain->GetSize() / 2));
-	m_Camera->LookAt(glm::vec3(m_Terrain->GetSize() / 2, 180.0f, m_Terrain->GetSize() / 2));
-	//m_Render = std::make_unique<TriangleRenderer>(*m_Terrain);
-	m_Render = std::make_shared<Geomipmapping>(*m_Terrain, 33.0f);
-	Input::DisableCursor();
+
 }
 
 void GameScene::Update(float deltaTime)
@@ -65,11 +46,10 @@ void GameScene::Render()
 	shader.SetMat4("view", view);
 
 	shader.SetVec3("reverseLightDir", m_Sun->GetReverseLightDirection());
-	m_Render->Render(shader, m_Camera->GetPosition());
-
+	m_Renderer->Render(shader, m_Camera->GetPosition());
 }
 
 GameScene::~GameScene()
 {
-	m_Terrain->UnloadHeightMap();
+	
 }